refactor(http): package 정리 및 ContentType 책임 분배
