+++ src/main/java/http/common/HttpHeader.java
@@ -1,10 +1,11 @@
-package http;
+package http.common;
 
 public enum HttpHeader {
     CONTENT_TYPE("Content-Type"),
     SET_COOKIE("Set-Cookie"),
     CONTENT_LENGTH("Content-Length"),
-    LOCATION("Location");
+    LOCATION("Location"),
+    COOKIE("Cookie");
 
     private final String text;
