+++ src/main/java/http/request/Uri.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 
 import utils.StringUtils;
