+++ src/main/java/http/response/ContentType.java
@@ -1,6 +1,4 @@
-package http;
-
-import java.util.EnumSet;
+package http.response;
 
 public enum ContentType {
 
@@ -22,24 +20,15 @@ public enum ContentType {
     public static final String CHARSET_UTF_8 = "charset=utf-8";
     private static final String SEMI_COLON = ";";
 
-    public static final String TEMPLATE_PATH = "./templates";
-    public static final String STATIC_PATH = "./static";
-    private static final EnumSet<ContentType> typesInTemplates = EnumSet.of(ContentType.HTML, ContentType.ICO);
-
     ContentType(String mediaType) {
         this.mediaType = mediaType;
     }
 
-    public static ContentType from(Uri uri) {
-        return valueOf(uri.getExtension().toUpperCase());
+    public static ContentType from(String extension) {
+        return valueOf(extension.toUpperCase());
     }
 
     public String toStringWithCharsetUTF8() {
         return this.mediaType + SEMI_COLON + CHARSET_UTF_8;
-    }
-
-    public static String toRelativePath(Uri uri) {
-        String prefix = typesInTemplates.contains(from(uri)) ? TEMPLATE_PATH : STATIC_PATH;
-        return prefix + uri.getPath();
     }
 }
