+++ src/test/java/http/HttpMethodTest.java
@@ -1,5 +1,6 @@
 package http;
 
+import http.request.HttpMethod;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
