+++ src/test/java/http/HttpProtocolTest.java
@@ -1,5 +1,6 @@
 package http;
 
+import http.request.HttpProtocol;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
