+++ src/test/java/http/RequestLineTest.java
@@ -1,5 +1,6 @@
 package http;
 
+import http.request.RequestLine;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
