+++ src/test/java/http/UriTest.java
@@ -1,5 +1,7 @@
 package http;
 
+import http.request.QueryString;
+import http.request.Uri;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
